Target uploads to new dropbox address
SSH server which was used to receive file uploads (CI, nightly and release) has been
migrated to a new one on a different address.
Username, Key and paths are unchanged.
Most notable changes are the use of master.download.kiwix.org as the target in
replacement of mirror.download.kiwix.org (although it would still work) and the
Port to which SSH is listening on (30022 instead of 22)

I changed the expected `remote` file format: must include a port number (or `::`).
You'll notice I changed the logic a bit so that top/customs/projects ZIP files are
extracted on disk first then uploaded.
Reason behind this is that there is no shell on that server so you can't unzip on the remote anymore.
I've looked at files and I don't think file size is an issue in that case.