Move to new linktarget table
Since Mediawiki 1.38, page links targets are stored in a new
`linktarget` table instead of the original `pagelinks` table.

Details:
- https://www.mediawiki.org/wiki/Manual:Linktarget_table
- https://www.mediawiki.org/wiki/Manual:Pagelinks_table

This commit modifies the query on pagelinks to retrieve link target
title from the new table instead of the former one.

This is a bit quick and dirty because this is not supposed to be used in
the long term since this tool is supposed to be decommissioned /
replaced by a Python alternative.